requirements.txt
fastavro==0.14.5
kafka-python==1.3.3
leveldb==0.194
lz4==0.10.1
psutil==5.2.2
python-snappy==0.5.1
PyYAML==3.12
setproctitle==1.1.10